video_quality/requirements_worldarena_vlm.txt
### Slim VLM requirements

triton==3.5.1
sentencepiece==0.2.1
transformers>=4.57.0,<4.58.0
qwen-vl-utils==0.0.14
huggingface_hub==0.36.2
safetensors==0.7.0
decord==0.6.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
imageio==2.31.1
imageio-ffmpeg==0.6.0
moviepy==1.0.3
einops==0.8.1
einops-exts==0.0.4
ftfy==6.3.1
open_clip_torch==3.2.0
pycocotools==2.0.11
scikit-learn==1.7.2
scipy==1.15.3
pandas==2.3.3
numpy==1.26.4
tqdm==4.67.1
accelerate==1.12.0
bitsandbytes==0.49.0
gdown==5.2.0
requests==2.32.5
omegaconf==2.3.0
python-dotenv==1.2.1
fsspec==2025.10.0
pytest==7.2.0
ipywidgets>=8.1.5
PyYAML==6.0.3
Pillow==10.0.0
